amplifi-dev/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["jonra1993 <jon_ra@hotmail.es>"]

[tool.black]
line-length = 88
target-version = ["py310", "py311"]
exclude = "((.eggs | .git | .pytest_cache | build | dist))"

[tool.isort]
profile = "black"
line_length = 88
skip = [".venv", "build", "dist"]

[tool.ruff]
line-length = 88
exclude = [".git", "__pycache__", ".mypy_cache", ".pytest_cache"]
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    # "I",  # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
lint.ignore = [
    "B904",
    "B006",
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.bandit]
exclude_dirs = ["test"]
skips = ["B104"]
disable = ["B403", "B301"] # Ignore pkl errors for now

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["alembic", "__pycache__"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
alembic = "^1.13.0"
asyncpg = "^0.29.0"
fastapi = { extras = ["all"], version = "^0.116.1" }
sqlmodel = "0.0.16"
cryptography = "^43.0.1"
bcrypt = "^4.1.1"
SQLAlchemy-Utils = "^0.41.1"
SQLAlchemy = "^2.0.23"
fastapi-pagination = { extras = ["sqlalchemy"], version = "^0.12.21" }
fastapi-cache2 = "^0.2.1"
watchfiles = "^0.21.0"
asyncer = "0.0.5"
httpx = "^0.28.1"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
redis = "5.0.4"
fastapi-async-sqlalchemy = "^0.6.0"
celery = "^5.3.6"
psycopg2-binary = "^2.9.5"
#transformers = "^4.35.2"
#torch = [
#    { url = "https://download.pytorch.org/whl/cpu/torch-2.0.0%2Bcpu-cp311-cp311-linux_x86_64.whl", markers = "sys_platform == 'linux'" },
#    { url = "https://download.pytorch.org/whl/cpu/torch-2.0.0%2Bcpu-cp311-cp311-win_amd64.whl", markers = "sys_platform == 'win32'" }
#]
requests = "^2.31.0"
wheel = "^0.42.0"
setuptools = "^78.1.1"
#langchain = "^0.0.348"
#openai = "^1.3.8"
fastapi-limiter = "^0.1.5"
pyjwt = { extras = ["crypto"], version = "^2.8.0" }
uvicorn = ">=0.27.0.post1, <0.28.0"
tzdata = "^2023.3"
#unstructured = {extras = ["pdf"], version = "^0.15.9"}
python-magic = "^0.4.27"
opencv-python = "^4.10.0.84"
azure-cognitiveservices-speech = "^1.35.0"
pydub = "^0.25.1"

# r2r dependency removed as it's no longer required
#r2r = {git = "https://github.com/thoughtswinsystems/R2R.git", rev = "TW-v2.0.0", subdirectory = "py"}
#r2r = {git = "https://github.com/thoughtswinsystems/R2R.git", rev = "TW-v1.5.0", subdirectory = "py"}

toml = "^0.10.2"
instructor = "^1.4.3"
gevent = "^24.2.1"
hvac = "^2.3.0"
tiktoken = "^0.8.0"
opencensus-ext-azure = "^1.1.13"
azure-monitor-opentelemetry = "^1.6.2"
croniter = "^3.0.3"
aiofiles = "^24.1.0"
loginradius-v2 = "^11.6.0"
smart-open = "^7.0.5"
azure-storage-blob = "^12.24.0"
tenacity = "^9.0.0"
sqlalchemy-celery-beat = "^0.8.0"
azure-identity = "^1.19.0"
azure-keyvault-secrets = "^4.9.0"
pdfplumber = "^0.11.5"
Pillow = "^10.0.0"
numpy = "^1.24.0"

kaleido = "0.2.1"
vanna = {extras = ["openai", "pgvector", "postgres"], version = "^0.7.5"}
psycopg = {extras = ["binary", "pool"], version = "^3.2.4"}
langdetect = "^1.0.9"
databricks-sql-connector = "^4.0.0"
markdown = "^3.7"
pypdf = "^3.9.0"
python-docx = "^1.1.2"
bs4 = "^0.0.2"
limits = "^4.6"
defusedxml = "^0.7.1"
pytesseract = "^0.3.13"
pgvector = "<0.4.0"
django = "^5.2"
tavily-python = "^0.7.2"
pydantic-ai = "^1.0.4"
logfire = "^3.14.1"
pymupdf = "^1.25.5"
reportlab = "^4.4.1"
isort = "^6.0.1"
pyarmor = { version = "^9.1.7", optional = true }
unstructured = "^0.17.2"
fitz = "^0.0.1.dev2"
aiomysql = "^0.2.0"
pymysql = "^1.1.1"
python-pptx = "^1.0.2"
uv = "^0.8.6"
pycryptodome = "^3.23.0"
kuzu = "^0.11.0"
rapidfuzz = "^3.13.0"
openai = "^1.99.8"
flagsmith = "^3.10.0"
chardet = "^5.2.0"
aioboto3 = "^15.1.0"
sqlean-py = "3.49.1"
starlette = "^0.47.2"
scikit-learn = "^1.7.2"
# Video processing dependencies
moviepy = "^1.0.3"
faster-whisper = "^1.0.3"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ruff = "^0.8.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-json = "^0.4.0"
pytest-json-ctrf = "^0.3.5"
pytest-cov = "^5.0.0"
mypy = "^1.5.0"
bandit = "^1.7.10"
radon = "^6.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
